refactor(dashboard_screen): implement transactions view
